improve config models modules
